refactor(stella-cli): move export.rs's test module into export/tests.rs
The fix in the previous commit took export.rs to 1482 of the 1500-line
god-file ceiling. The ratchet judges a change against what it inherited, so
that is not a failure today — it is a landmine for whoever adds the next test,
who would fail `make file-size` for a reason they did not cause.

Same shape as `agent.rs` + `agent/tests.rs`. Pure move: every test keeps its
name and its `export::tests::*` path, so nothing reads as a deletion and no
baseline entry is needed (878 and 610 lines, both well under 1500).

Refs #2558
